pyproject.toml
[tool.flake8]
max-line-length = 88
exclude = "*/migrations/*, requirements.txt, requirements.dev.txt, swagger.yml"

[tool.isort]
profile = "black"
multi_line_output = 3
ignore_init_module_imports = true


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings.test"
addopts = "-p no:warnings"
env = [{ ENVIRONMENT = "test" }]
pytest_plugins = "utils.fixtures"


python_files = "tests/* test/*"
norecursedirs = "__IGNORE/*"
[tool.black]
include = '\.py$'
line-length = 88

[tool.poetry]
name = "backend"
version = "2.3.3"
description = ""
authors = ["rkesik <rafal.kesik@gmail.com>", "bartnyk <jakub@playmaker.pro>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.10,<3.9"
Django = "^3.2"
django-fsm = "2.7.1"
django-countries = "7.1"
django-cities-light = "3.9.2"
django-phonenumbers = "1.0.1"
django-extensions = "3.2.3"
easy-thumbnails = "^2.8.5"
django-user-agents = "^0.4.0"
django-phonenumber-field = "5.0.0"
django-cleanup = "^8.0.0"
django-address = "0.2.5"
djangorestframework = "3.12.1"
djangorestframework-simplejwt = "5.2.2"
django-cors-headers = "3.4.0"
django-allauth = "0.50.0"
drf-spectacular = "^0.26.4"
drf-yasg = "^1.21.7"
drf-spectacular-sidecar = "^2023.9.1"
django-select2 = "^7.10"
pm-core = { version = "^1.0.0", source = "pm" }
django-filter = "^23.2"
psycopg2-binary = "^2.9.7"
pygments = "^2.1"
django-crispy-forms = "^1.9"
python-dateutil = "^2.8.2"
pymemcache = "^4.0.0"
ipython = "^8.12.3"
sentry-sdk = "^1.39.1"
celery = "^5.4.0"
redis = "^5.2.1"
djangorestframework-yaml = "^2.0.0"
django-redis = "^5.4.0"
django-celery-beat = "^2.8.0"
dotenv = "^0.9.9"
pymongo = "4.7.3"
mongoengine = "0.27.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "2.2"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.7"
pytest-mock = "*"
django-test-migrations = "1.3.0"
factory-boy = "3.2.1"
pre-commit = "3.3.1"
parameterized = "^0.9.0"
black = "^23.9.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mongomock = "^4.3.0"


[[tool.poetry.source]]
name = "pm"
url = "http://jacekplaymaker.usermd.net:56258/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
extend-select = ["I", "E4", "E7", "E9", "F"]
ignore = ["E731", "E741", "F821", "B904", "E722", "F405", "F403"]
ignore-init-module-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]